May 14/requirements.txt
PyPDF2==3.0.1
chromadb==0.4.22
google-generativeai==0.3.2
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.17
